+++ notification-center/src/Dignite.Abp.NotificationCenter.MongoDB/Dignite/Abp/NotificationCenter/MongoDB/AbpNotificationCenterMongoDbModule.cs
@@ -16,5 +16,7 @@ public class AbpNotificationCenterMongoDbModule : AbpModule
         {
             options.AddDefaultRepositories(includeAllEntities: true);
         });
+
+        context.Services.AddHostedService<NotificationCenterMongoDbOutboxCapabilityHostedService>();
     }
 }